Migrate to Spark 2.0
