+++ src/main/scala/com/databricks/spark/xml/util/XmlFile.scala
@@ -15,12 +15,20 @@
  */
 package com.databricks.spark.xml.util
 
+import java.io.CharArrayWriter
 import java.nio.charset.Charset
+import javax.xml.stream.XMLOutputFactory
 
+
+import scala.collection.Map
+
+import com.databricks.spark.xml.parsers.StaxXmlGenerator
+import com.sun.xml.internal.txw2.output.IndentingXMLStreamWriter
 import org.apache.hadoop.io.{Text, LongWritable}
 
-import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
+import org.apache.spark.SparkContext
+import org.apache.spark.sql.DataFrame
 import com.databricks.spark.xml.{XmlOptions, XmlInputFormat}
 
 private[xml] object XmlFile {
@@ -45,6 +53,92 @@ private[xml] object XmlFile {
         classOf[XmlInputFormat],
         classOf[LongWritable],
         classOf[Text]).map(pair => new String(pair._2.getBytes, 0, pair._2.getLength, charset))
+    }
+  }
+
+  /**
+   * Note that writing a XML file from [[DataFrame]] having a field
+   * [[org.apache.spark.sql.types.ArrayType]] with its element as nested array would have
+   * an additional nested field for the element. For example, the [[DataFrame]] having
+   * a field below,
+   *
+   *   fieldA Array(Array(data1, data2))
+   *
+   * would produce a XML file below.
+   *
+   * <fieldA>
+   *     <item>data1</item>
+   * </fieldA>
+   * <fieldA>
+   *     <item>data2</item>
+   * </fieldA>
+   *
+   * Namely, roundtrip in writing and reading can end up in different schema structure.
+   */
+  def saveAsXmlFile(
+      dataFrame: DataFrame,
+      path: String,
+      parameters: Map[String, String] = Map()): Unit = {
+    val options = XmlOptions(parameters.toMap)
+    val codecClass = CompressionCodecs.getCodecClass(options.codec)
+    val startElement = s"<${options.rootTag}>"
+    val endElement = s"</${options.rootTag}>"
+    val rowSchema = dataFrame.schema
+    val indent = XmlFile.DEFAULT_INDENT
+    val rowSeparator = XmlFile.DEFAULT_ROW_SEPARATOR
+
+    val xmlRDD = dataFrame.rdd.mapPartitions { iter =>
+      val factory = XMLOutputFactory.newInstance()
+      val writer = new CharArrayWriter()
+      val xmlWriter = factory.createXMLStreamWriter(writer)
+      val indentingXmlWriter = new IndentingXMLStreamWriter(xmlWriter)
+      indentingXmlWriter.setIndentStep(indent)
+
+      new Iterator[String] {
+        var firstRow: Boolean = true
+        var lastRow: Boolean = true
+
+        override def hasNext: Boolean = iter.hasNext || firstRow || lastRow
+
+        override def next: String = {
+          if (iter.nonEmpty) {
+            val xml = {
+              StaxXmlGenerator(
+                rowSchema,
+                indentingXmlWriter,
+                options)(iter.next())
+              writer.toString
+            }
+            writer.reset()
+
+            // Here it needs to add indentations for the start of each line,
+            // in order to insert the start element and end element.
+            val indentedXml = indent + xml.replaceAll(rowSeparator, rowSeparator + indent)
+            if (firstRow) {
+              firstRow = false
+              startElement + rowSeparator + indentedXml
+            } else {
+              indentedXml
+            }
+          } else {
+            indentingXmlWriter.close()
+            if (!firstRow) {
+              lastRow = false
+              endElement
+            } else {
+              // This means the iterator was initially empty.
+              firstRow = false
+              lastRow = false
+              ""
+            }
+          }
+        }
+      }
+    }
+
+    codecClass match {
+      case null => xmlRDD.saveAsTextFile(path)
+      case codec => xmlRDD.saveAsTextFile(path, codec)
     }
   }
 }
